* Reverted the getMagicLevel and fixed up as it should be * Whitespaces on tools
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@3381 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -2101,7 +2101,7 @@ int32_t LuaScriptInterface::luaCanPlayerLearnInstantSpell(lua_State* L)
 			return 1;
 		}
 
-		if(player->getMagicLevel() < spell->getMagicLevel())
+		if((int32_t)player->getMagicLevel() < spell->getMagicLevel())
 		{
 			lua_pushnumber(L, LUA_FALSE);
 			return 1;
